Documents/Gradio-Dashboard/requirements.txt
gradio
torch
ultralytics
sahi
deep_sort_realtime
opencv-python
pillow
